requirements.txt
streamlit>=1.30.0
googlesearch-python>=1.3.0
